Plugins: Refactor call resource API handling
